Update Python-CA
Added the functionality of asking user to add more input of student records in the already sorted list of student records. If user selects Yes the new record is added at the end of the sorted list and Insertion sort is used to sort the new list as it performs best in case of adding element in a sorted list or array and lowest student record is once again returned. Once User selects no program exits giving the count of total elements in the list